+++ webapp/src/App.tsx
@@ -12,16 +12,16 @@ import ValidacionEmail from "./vistas/registroLogin/ValidacionEmail";
 import UserHistory from "./vistas/UserHistory";
 import VariantSelect from "./vistas/VariantSelect";
 import MultiplayerLobby from "./vistas/MultiplayerLobby";
-import GameMultiplayer from "./vistas/GameMultiplayer";
-import GameFortuneCoin from "./vistas/GameFortuneCoin";
-import GameFortuneDice from "./vistas/GameFortuneDice";
-import GameHoley from "./vistas/GameHoley";
-import GameMaster from "./vistas/GameMaster";
-import GamePastel from "./vistas/GamePastel";
-import GamePolyY from "./vistas/GamePolyY";
-import GameTabu from "./vistas/GameTabu";
-import GameWhyNot from "./vistas/GameWhyNot";
-import Game3DY from "./vistas/Game3DY";
+import GameMultiplayer from "./vistas/game/GameMultiplayer";
+import GameFortuneCoin from "./vistas/game/GameFortuneCoin";
+import GameFortuneDice from "./vistas/game/GameFortuneDice";
+import GameHoley from "./vistas/game/GameHoley";
+import GameMaster from "./vistas/game/GameMaster";
+import GamePastel from "./vistas/game/GamePastel";
+import GamePolyY from "./vistas/game/GamePolyY";
+import GameTabu from "./vistas/game/GameTabu";
+import GameWhyNot from "./vistas/game/GameWhyNot";
+import Game3DY from "./vistas/game/Game3DY";
 import GameHvB from "./vistas/game/GameHvB";
 import GameHvH from "./vistas/game/GameHvH";
 import {
